feat: add support for Slurm 20.11.7-1
